python 2 compatibility moved print_conditions from neq into RADIS, update code accordingly create a spectrum.utils file with most definitions
